% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRBfact.R
\docType{data}
\name{LRBfact}
\alias{LRBfact}
\title{Binford's data - labeled version}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 339 rows and 507 columns.}
\source{
\url{http://ajohnson.sites.truman.edu/data-and-program/}
}
\usage{
LRBfact
}
\description{
Datasets used to calculate Binford's environmental and hunter-gatherer frames of reference variables used in Binford (2001). LRBfact is equal to LRB except code values are represented by their label, rather than the code.
}
\details{
Binford, Lewis R. 2001. Constructing Frames of Reference: An Analytical Method for Archaeological Theory Building Using Ethnographic and Environmental Data Sets University of California Press, Berkeley

The key to the variables in this object is in LBRkey

Downloaded from the source url on 31 Jul 2016
}
\keyword{datasets}
