% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRBkey.R
\docType{data}
\name{LRBkey}
\alias{LRBkey}
\title{Key to Binford's data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 506 rows and 16 columns.}
\source{
\url{http://ajohnson.sites.truman.edu/data-and-program/}
}
\usage{
LRBkey
}
\description{
Data frame describing the variables in the LRB data frame. Part of the datasets used to calculate Binford's environmental and hunter-gatherer frames of reference variables used in Binford (2001).
}
\details{
Binford, Lewis R. 2001. Constructing Frames of Reference: An Analytical Method for Archaeological Theory Building Using Ethnographic and Environmental Data Sets University of California Press, Berkeley

Downloaded from the source url on 31 Jul 2016
}
\keyword{datasets}
