% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRB.R
\docType{data}
\name{LRB}
\alias{LRB}
\title{Binford's data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 339 rows and 507 columns.}
\source{
\url{http://ajohnson.sites.truman.edu/data-and-program/}
}
\usage{
LRB
}
\description{
Datasets used to calculate Binford's environmental and hunter-gatherer frames of reference variables used in:
}
\details{
Binford, Lewis R. 2001. Constructing Frames of Reference: An Analytical Method for Archaeological Theory Building Using Ethnographic and Environmental Data Sets University of California Press, Berkeley

Key to variables in
 LBRkey.csv
 LRBcodebook.txt

Downloaded from the source url on 31 Jul 2016
}
\keyword{datasets}

